// idcs-multi-user-kmsi/server/types/idcsTypes.d.ts
interface IdcsResourceList {
  Resources: Array<unknown>;
  totalResults: number;
  startIndex?:number;
  itemsPerPage?:number;
}

//Users obviously have lots of fields, so this can be fleshed out as required
interface User {
  id: string;
  active?: boolean;
  "urn:ietf:params:scim:schemas:oracle:idcs:extension:passwordState:User"?: UserPasswordState
}

interface UserPasswordState {
  applicablePasswordPolicy?: UserPasswordPolicy
}

interface UserPasswordPolicy {
  display?:string;
  priority:number;
  value:string;
}
// ...
interface TokenResponse {
  access_token: string;
  token_type: string;
  expires_in: number;
  refresh_token?: string;
  id_token?: string;
}
// ...
interface TrustedUserAgent {
  id:string;
  trustToken:string;
  name:string;
  user?:UserReference;
  expiryTime?:string;
  tokenType?:"KMSI"|"TRUSTED";
}

interface UserReference {
  value?:string;
  ocid?:string;
  "$ref"?:string;
}
// ...
interface CountryCodesAllowedValues {
  attrValues: Array<CountryCode>;
}

interface CountryCode {
  label: string;
  value: string;
}

interface PasswordPolicy {
  name?:string;
  id?:string;
  minLength?: number;
  maxLength?: number;
  minLowerCase?: number;
  minUpperCase?: number;
  minAlphas?: number
  minNumerals?: number;
  minAlphaNumerals?: number;
  minSpecialChars?: number;
  minUniqueChars?: number;
  maxRepeatedChars?: number;
  startsWithAlphabet?: boolean;
  userNameDisallowed?: boolean;
  firstNameDisallowed?: boolean;
  lastNameDisallowed?: boolean; 
  disallowedChars?: string;
  requiredChars?:string;
  numPasswordsInHistory?: number;
}